fix qep remove warning

+++ motorlib/config_g474_boost.cpp
@@ -13,7 +13,7 @@ static struct {
     GPIO motor_encoder_cs = {*GPIOA, 15, GPIO::OUTPUT};
     SPIEncoder motor_encoder = {*SPI3, motor_encoder_cs};
     GPIO enable = {*GPIOC, 11, GPIO::OUTPUT};
-    HRPWM motor_pwm = {initial_param.fast_loop_param.pwm_frequency, *HRTIM1, 3, 5, 4};
+    HRPWM motor_pwm = {static_cast<uint32_t>(initial_param.fast_loop_param.pwm_frequency), *HRTIM1, 3, 5, 4};
     FastLoop fast_loop = {motor_pwm, motor_encoder};
     LED led = {const_cast<uint16_t*>(reinterpret_cast<volatile uint16_t *>(&TIM4->CCR1)), 
                const_cast<uint16_t*>(reinterpret_cast<volatile uint16_t *>(&TIM4->CCR2)),
